Use latest SDK and Plugin builders
- Fix the plugin builders to the be the new one only
- Fix the SDK to use the angular client properly
- Adjusted a bit the seed templates to produce more complete project

Testing Done:
```shell
vcd-ext new test
cd test && npm run build
npm run serve
vcd-ext deploy -f --only=ui
```
