+++ packages/angular/package.json
@@ -1,29 +1,24 @@
 {
   "name": "@vcd/angular-sdk",
-  "version": "0.13.0-alpha.0",
+  "version": "0.13.0-alpha.1",
   "scripts": {
     "ng": "ng",
     "build": "npm run build:sdk && npm run build:schematics && npm run build:plugin-builders",
     "build:sdk": "ng build @vcd/sdk && npm run sdk:build:prod",
-    "build:plugin-builders": "npm run plugin-builders:install:dev:dep && ng build @vcd/plugin-builders && npm run plugin-builders:build",
+    "build:plugin-builders": "ng build @vcd/plugin-builders && npm run plugin-builders:build",
     "build:schematics": "ng build @vcd/schematics && npm run schematics:build:prod",
-
     "sdk:build": "tsc -p ./projects/vcd/sdk/tsconfig.schematics.json",
     "sdk:build:prod": "npm run sdk:build && npm run sdk:copy:schemas && npm run sdk:fix:backwards:compatability",
     "sdk:copy:schemas": "cp -R ./projects/vcd/sdk/src/schematics ./dist/vcd/sdk",
     "sdk:fix:backwards:compatability": "node ./scripts/update_packagejson.js",
-
     "plugin-builders:build:schematics": "tsc -p ./projects/vcd/plugin-builders/tsconfig.schematics.json",
     "plugin-builders:copy:schemas": "cp -R ./projects/vcd/plugin-builders/src/lib/schematics ./dist/vcd/plugin-builders",
     "plugin-builders:build": "tsc --build ./projects/vcd/plugin-builders/tsconfig.json && npm run plugin-builders:cpy:schema && npm run plugin-builders:cpy:builders && npm run plugin-builders:build:schematics && npm run plugin-builders:copy:schemas",
     "plugin-builders:cpy:schema": "cp ./projects/vcd/plugin-builders/src/lib/base/schema.json ./dist/vcd/plugin-builders",
     "plugin-builders:cpy:builders": "cp ./projects/vcd/plugin-builders/builders.json ./dist/vcd/plugin-builders",
-    "plugin-builders:install:dev:dep": "cd ./projects/vcd/plugin-builders && npm i && cd ./src/lib/new && npm i",
-
     "schematics:build": "tsc -p ./projects/vcd/schematics/tsconfig.json",
     "schematics:build:prod": "npm run schematics:build && npm run schematics:copy:schemas",
     "schematics:copy:schemas": "cp -R ./projects/vcd/schematics/src/schematics ./dist/vcd/schematics",
-
     "test": "ng test",
     "test-ci": "ng test --configuration=test-ci",
     "lint": "ng lint"
@@ -40,6 +35,7 @@
     "@angular/router": "8.2.14",
     "@ngrx/core": "1.2.0",
     "@ngrx/store": "9.1.0",
+    "@vcd/angular-client": "file:../../../cloud-director-typescript-clients/packages/angular/dist/vcd-angular-client-0.0.2-alpha.1.tgz",
     "@vcd/bindings": "9.1.1",
     "rxjs": "6.4.0",
     "rxjs-compat": "6.5.5",
