+++ packages/angular/projects/vcd/plugin-builders/builders.json
@@ -1,9 +1,9 @@
 {
     "builders": {
         "plugin-builder": {
-            "class": "./base/index.js",
+            "implementation": "./base/index.js",
             "schema": "./schema.json",
-            "description": "Build vCloud Director UI Plugin for version 9.1 - 10.0"
+            "description": "Build VMware Cloud Director UI Plugin for version 10.1+"
         }
     }
 }
