+++ packages/angular/projects/vcd/plugin-builders/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@vcd/plugin-builders",
-  "version": "0.12.1-alpha.0",
+  "version": "0.12.1-alpha.1",
   "description": "Angular CLI-compatible tools for building Cloud Director plugins.",
   "author": "VMware",
   "license": "BSD-2-Clause",
@@ -10,7 +10,7 @@
     "zip-webpack-plugin": "3.0.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "0.8.0"
+    "@angular-devkit/build-angular": "0.803.29"
   },
   "peerDependencies": {
     "@angular-devkit/build-angular": "0.8.0",
