--- packages/angular/projects/vcd/plugin-builders/src/lib/new/package.json
@@ -1,8 +1,0 @@
-{
-    "dependencies": {
-        "zip-webpack-plugin": "3.0.0"
-    },
-    "devDependencies": {
-        "@angular-devkit/build-angular": "0.803.29"
-    }
-}
