+++ packages/angular/projects/vcd/sdk/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@vcd/sdk",
-  "version": "0.12.2-alpha.0",
+  "version": "0.12.2-alpha.1",
   "description": "Angular-based client for the Cloud Director API.",
   "author": "VMware",
   "license": "BSD-2-Clause",
@@ -12,12 +12,13 @@
   },
   "module": "esm5/vcd-sdk.js",
   "dependencies": {
-    "@vcd/angular-client": "0.0.2-alpha.0"
+    "@vcd/angular-client": "0.0.2-alpha.1"
   },
   "peerDependencies": {
     "@angular-devkit/schematics": "7.3.8",
     "@angular/common": "4.4.4 || 8.2.14",
     "@angular/core": "4.4.4 || 8.2.14",
+    "@vcd/angular-client": "0.0.2-alpha.1",
     "@ngrx/core": "1.2.0",
     "@ngrx/store": "2.2.2",
     "@schematics/angular": "9.1.4",
