+++ packages/angular/projects/vcd/sdk/src/public-api.ts
@@ -3,7 +3,6 @@
  */
 
 export * from './main';
-export * from '@vcd/angular-client/client';
 export * from './common';
 export * from './core';
-export * from '@vcd/angular-client/query';
+export * from '@vcd/angular-client';
