+++ packages/cli/templates/uiPlugin/package.json
@@ -27,7 +27,7 @@
     "@clr/ui": "3.1.3",
     "@ngrx/store": "6.0.1",
     "@ngtools/webpack": "9.1.7",
-    "@vcd/sdk": "0.12.2-alpha.0",
+    "@vcd/sdk": "0.12.2-alpha.1",
     "@vcd/i18n": "1.0.0-dev.1",
     "@vcd/bindings": "9.1.1",
     "@vcd/ui-components": "1.0.0-dev.46",
@@ -51,7 +51,7 @@
     "@types/jasminewd2": "2.0.3",
     "@types/node": "8.9.4",
     "@vcd/ext-cli": "0.0.12-alpha.2",
-    "@vcd/plugin-builders": "0.12.1-alpha.0",
+    "@vcd/plugin-builders": "0.12.1-alpha.1",
     "@vcd/ui-emulator": "0.0.6-alpha.1",
     "codelyzer": "4.3.0",
     "jasmine-core": "2.99.1",
